almost ready

+++ app/src/Controller/CategoryController.php
@@ -9,13 +9,13 @@ namespace App\Controller;
 use App\Entity\Category;
 use App\Service\CategoryServiceInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\MapQueryParameter;
 use Symfony\Component\Routing\Attribute\Route;
 use App\Form\Type\CategoryType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Contracts\Translation\TranslatorInterface;
-
 /**
  * Class CategoryController.
  */
@@ -151,7 +151,7 @@ class CategoryController extends AbstractController
         );
     }
 
-   /**
+       /**
      * Delete action.
      *
      * @param Request  $request  HTTP request
@@ -167,6 +167,15 @@ class CategoryController extends AbstractController
     )]
     public function delete(Request $request, Category $category): Response
     {
+        if (!$this->categoryService->canBeDeleted($category)) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.category_contains_tasks')
+            );
+
+            return $this->redirectToRoute('category_index');
+        }
+
         $form = $this->createForm(FormType::class, $category, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
@@ -191,5 +200,5 @@ class CategoryController extends AbstractController
                 'category' => $category,
             ]
         );
-    }  
+    }
 }
